Pre-process all ES5 builds (#2840)
