+++ packages/firestore/rollup.shared.js
@@ -21,6 +21,24 @@ import { externs } from './externs.json';
 import { renameInternals } from './scripts/rename-internals';
 import { extractPublicIdentifiers } from './scripts/extract-api';
 
+import pkg from './package.json';
+
+const browserDeps = Object.keys(
+  Object.assign({}, pkg.peerDependencies, pkg.dependencies)
+);
+
+const nodeDeps = [...browserDeps, 'util', 'path'];
+
+/** Resolves the external dependencies for the browser build. */
+export function resolveBrowserExterns(id) {
+  return browserDeps.some(dep => id === dep || id.startsWith(`${dep}/`));
+}
+
+/** Resolves the external dependencies for the Node build. */
+export function resolveNodeExterns(id) {
+  return nodeDeps.some(dep => id === dep || id.startsWith(`${dep}/`));
+}
+
 const externsPaths = externs.map(p => path.resolve(__dirname, '../../', p));
 const publicIdentifiers = extractPublicIdentifiers(externsPaths);
